Debug flag on gradle

+++ src/test/groovy/com/example/rpl/RPL/controller/CoursesControllerFunctionalSpec.groovy
@@ -1,23 +1,7 @@
 package com.example.rpl.RPL.controller
 
-import com.example.rpl.RPL.model.Activity
-import com.example.rpl.RPL.model.ActivityCategory
-import com.example.rpl.RPL.model.ActivitySubmission
-import com.example.rpl.RPL.model.Course
-import com.example.rpl.RPL.model.CourseUser
-import com.example.rpl.RPL.model.Language
-import com.example.rpl.RPL.model.RPLFile
-import com.example.rpl.RPL.model.Role
-import com.example.rpl.RPL.model.SubmissionStatus
-import com.example.rpl.RPL.model.User
-import com.example.rpl.RPL.repository.ActivityCategoryRepository
-import com.example.rpl.RPL.repository.ActivityRepository
-import com.example.rpl.RPL.repository.CourseRepository
-import com.example.rpl.RPL.repository.CourseUserRepository
-import com.example.rpl.RPL.repository.FileRepository
-import com.example.rpl.RPL.repository.RoleRepository
-import com.example.rpl.RPL.repository.SubmissionRepository
-import com.example.rpl.RPL.repository.UserRepository
+import com.example.rpl.RPL.model.*
+import com.example.rpl.RPL.repository.*
 import com.example.rpl.RPL.util.AbstractFunctionalSpec
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.crypto.password.PasswordEncoder
@@ -204,7 +188,7 @@ class CoursesControllerFunctionalSpec extends AbstractFunctionalSpec {
             assert course.description == body.description
             assert course.semester == body.semester
 
-            assert courseUserRepository.findByCourse_IdAndUser_Id(course.id, user.getId()).isPresent()
+            assert courseUserRepository.findByCourse_IdAndUser_Id(course.id as Long, user.getId() as Long).isPresent()
 
             assert courseRepository.existsById(course.id as Long)
     }
